create a chuck norris pod/domain and move every component inside
The api is still local to the pod and separated from the mappers and viewmodel, but 'global' to the components inside the pod.

And being separated in a folder, makes everything more legible. 

Also, all the containers have been moved to component folder, and inside each container, they have their own subcomponents. So everything is tidy and easy to find.